Add the output bit width into the name of the UF that give TC coefficients
Uninterpreted functions with different input signatures should be
regarded as different even if they have the same name.  To avoid
this confusion and as a general point of style, this commit adds the
width to the name of the uninterpreted functions used for the
coefficients of the Tooms-Cook result polynomial.